% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCIs_correlated_ANOVA.R
\name{SCI_ANOVA_Toep}
\alias{SCI_ANOVA_Toep}
\title{Title Function to evaluate the simultaneous confidence intervals of the pairwise differences of the mean effects}
\usage{
SCI_ANOVA_Toep(data, size, t, k, n)
}
\arguments{
\item{data}{A numeric data frame consists of cell data sets in the one-way ANOVA}

\item{size}{The level of significance}

\item{t}{The number of time (space) points}

\item{k}{The number of groups}

\item{n}{A numeric vector consists of cell sample sizes in the one-way ANOVA}
}
\value{
Simultaneous confidence intervals of the pairwise differences of the mean treatment effects
}
\description{
Title Function to evaluate the simultaneous confidence intervals of the pairwise differences of the mean effects
}
\examples{
data<-read.csv("https://raw.githubusercontent.com/RajuProbStat/OnewayCorrToep/refs/heads/main/DBP_data.csv")
SCI_ANOVA_Toep(data,0.05,3,3,c(6,6,8))
}
