% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_special_Toeplitz_structure.R
\name{T_1_T_1_A}
\alias{T_1_T_1_A}
\title{Function to evaluate the observed, critical and P-values of likelihood ratio and asymptotic likelihood ratio tests for testing covariance structure}
\usage{
T_1_T_1_A(data, size, t, n)
}
\arguments{
\item{data}{A data frame of dimension t by n consists of n number of t-dimensional cell data}

\item{size}{The level of significance, say 0.05}

\item{t}{The number of time (space) points, say t}

\item{n}{Sample size, say n}
}
\value{
Numeric values consist of observed, critical and P-values of the likelihood ratio and asymptotic likelihood ratio tests
}
\description{
Function to evaluate the observed, critical and P-values of likelihood ratio and asymptotic likelihood ratio tests for testing covariance structure
}
\examples{
data<-read.csv("https://raw.githubusercontent.com/RajuProbStat/OnewayCorrToep/refs/heads/main/DBP_data.csv")
T_1_T_1_A(data[1:6],0.05,3,6)
data<-read.csv("https://raw.githubusercontent.com/RajuProbStat/OnewayCorrToep/refs/heads/main/DBP_data.csv")
T_1_T_1_A(data[7:12],0.05,3,6)
data<-read.csv("https://raw.githubusercontent.com/RajuProbStat/OnewayCorrToep/refs/heads/main/DBP_data.csv")
T_1_T_1_A(data[13:20],0.05,3,8)
}
