% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_correlated_ANOVA.R
\name{T_2_T_2_A}
\alias{T_2_T_2_A}
\title{Function to obtain the observed, critical and P-values of the likelihood ratio and asymptotic likelihood ratio tests for testing homogeneity of mean effects in a one-way correlated ANOVA with a special Toeplitz error structure}
\usage{
T_2_T_2_A(data, size, t, k, n)
}
\arguments{
\item{data}{A numeric data frame consists of cell data sets in the one-way ANOVA}

\item{size}{The level of significance}

\item{t}{The number of time (space) points}

\item{k}{The number of groups}

\item{n}{A numeric vector consists of cell sample sizes in the one-way ANOVA}
}
\value{
Numeric values consist of observed, critical and P-values of the likelihood ratio and asymptotic likelihood ratio tests
}
\description{
Function to obtain the observed, critical and P-values of the likelihood ratio and asymptotic likelihood ratio tests for testing homogeneity of mean effects in a one-way correlated ANOVA with a special Toeplitz error structure
}
\examples{
data<-read.csv("https://raw.githubusercontent.com/RajuProbStat/OnewayCorrToep/refs/heads/main/DBP_data.csv")
T_2_T_2_A(data,0.05,3,3,c(6,6,8))
}
